BaseTheme is no more 🦀
